replace epk proof with report nullifier proof
